background_agent/celery_app.py
from celery import Celery
import os

REDIS_URL = os.environ.get('REDIS_URL', 'redis://localhost:6379/0')

celery_app = Celery('background_agent', broker=REDIS_URL, backend=REDIS_URL)


@celery_app.task(bind=True)
def process_task(self, task_id, agent, command, meta=None):
    # Placeholder for delegated processing via Celery
    return {'task_id': task_id, 'agent': agent, 'status': 'delegated'}
